refactor(favorite): move logic favorite delete into favorite structure
