+++ itachallenge-user/src/main/java/com/itachallenge/userinteraction/service/favorite/FavoriteService.java
@@ -9,4 +9,6 @@ public interface FavoriteService {
     Mono<Boolean> addChallengeToFavorites(String userId, String challengeId);
 
     Mono<Set<UUID>> getUserFavorites(String userId);
+
+    Mono<Boolean> deleteChallengeFromFavorites(String userId, String challengeId);
 }
